requirements.txt
openai>=1.13.3
pyautogui>=0.9.54
keyboard>=0.13.5
pillow>=10.2.0
python-dotenv>=1.0.1